requirements.txt
numpy==1.23.3
pandas==1.4.4
pulsar-client==2.10.1
